+++ gene_trajectory/add_gene_bin_score.py
