+++ gene_trajectory/coarse_grain.py
