+++ test/test_add_gene_bin_score.py
@@ -2,7 +2,7 @@ import unittest
 
 import numpy as np
 
-from gene_trajectories.add_gene_bin_score import add_gene_bin_score
+from gene_trajectory.add_gene_bin_score import add_gene_bin_score
 from test.example_data import example_adata, diffusion_map, gene_trajectories, to_adata, gene_expression
