+++ test/test_extract_gene_trajectory.py
@@ -3,7 +3,7 @@ import unittest
 import numpy as np
 import pandas as pd
 
-from gene_trajectories.extract_gene_trajectory import get_gene_embedding, get_randow_walk_matrix, get_gene_pseudoorder, \
+from gene_trajectory.extract_gene_trajectory import get_gene_embedding, get_randow_walk_matrix, get_gene_pseudoorder, \
     extract_gene_trajectory
 from test.example_data import gene_names, gene_trajectories
