+++ test/test_run_dm.py
@@ -2,7 +2,7 @@ import unittest
 
 import numpy as np
 
-from gene_trajectories.run_dm import run_dm
+from gene_trajectory.run_dm import run_dm
 from test.example_data import example_adata, diffusion_map
